refactor: Move network utilities to ui/core/utilities
Move api_request and internet_connection_checker from data/network
to ui/core/utilities and update all imports.
