+++ lib/data/auth/network/auth_api.dart
@@ -1,9 +1,9 @@
 import 'package:sport_matcher/data/auth/domain/auth_response.dart';
 import 'package:sport_matcher/data/auth/network/abstract_auth_api.dart';
 import 'package:sport_matcher/data/auth/network/email_registration_request.dart';
-import 'package:sport_matcher/data/network/api_request.dart';
-import 'package:sport_matcher/data/network/api_result.dart';
-import 'package:sport_matcher/data/network/http_method.dart';
+import 'package:sport_matcher/ui/core/utilities/api_request/api_request.dart';
+import 'package:sport_matcher/ui/core/utilities/api_request/api_result.dart';
+import 'package:sport_matcher/ui/core/utilities/api_request/http_method.dart';
 
 class AuthApi extends AbstractAuthApi {
   @override
