+++ lib/ui/authentication/sign_up/widgets/sign_up_screen_model.dart
@@ -1,6 +1,6 @@
 import 'package:sport_matcher/data/auth/network/auth_api.dart';
 import 'package:sport_matcher/data/auth/network/email_registration_request.dart';
-import 'package:sport_matcher/data/network/api_result.dart';
+import 'package:sport_matcher/ui/core/utilities/api_request/api_result.dart';
 
 class SignUpScreenModel {
   final AuthApi _authApi;
