+++ lib/ui/core/utilities/api_request/api_result.dart
