+++ lib/ui/core/utilities/internet_connection_checker/internet_connection_checker.dart
@@ -1,6 +1,6 @@
 import 'dart:io';
 
-import 'package:sport_matcher/data/network/abstract_internet_connection_checker.dart';
+import 'package:sport_matcher/ui/core/utilities/internet_connection_checker/abstract_internet_connection_checker.dart';
 
 class InternetConnectionChecker implements AbstractInternetConnectionChecker {
   @override
